pyproject.toml
[project]
name = "mcp-search-linkup"
version = "0.1.3"
description = "A MCP server to search web with Linkup"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "linkup-sdk>=0.2.0",
 "mcp>=1.0.0",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-search-linkup = "mcp_search_linkup:main"

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "S", # flake8-bandit
]

[dependency-groups]
dev = [
    "ruff>=0.8.2",
]
